+++ tests/integration_test.rs
@@ -1,0 +1,229 @@
+use assert_cmd::prelude::*; // Add methods on commands
+use std::str;
+use std::fs;
+use std::path::Path;
+use serial_test::serial;
+use std::process::Command; // Run programs
+
+fn fresh(){
+    Command::new("rm")
+        .arg("-r")
+        .args(["./tests/results/test_sketch_dir"])
+        .spawn();
+}
+
+#[serial]
+#[test]
+fn test_profile_vs_contain(){
+
+    let mut output = Command::cargo_bin("sylph").unwrap();
+    let output = output
+        .arg("profile")
+        .arg("./test_files/o157_reads.fastq")
+        .arg("./test_files/e.coli-EC590.fasta")
+        .output()
+        .expect("Output failed");
+    let stdout = str::from_utf8(&output.stdout).expect("Output was not valid UTF-8");
+    dbg!(stdout.matches('\n').count());
+    assert!(stdout.matches('\n').count() == 2);
+
+    let mut output = Command::cargo_bin("sylph").unwrap();
+    let output = output
+        .arg("contain")
+        .arg("./test_files/o157_reads.fastq")
+        .arg("./test_files/e.coli-EC590.fasta")
+        .arg("./test_files/e.coli-o157.fasta")
+        .arg("./test_files/e.coli-K12.fasta")
+        .output()
+        .expect("Output failed");
+    let stdout = str::from_utf8(&output.stdout).expect("Output was not valid UTF-8");
+    dbg!(stdout.matches('\n').count());
+    println!("{}",stdout);
+    assert!(stdout.matches('\n').count() == 4);
+}
+
+#[serial]
+#[test]
+fn test_sketch_commands() {
+    Command::new("rm")
+        .arg("-r")
+        .args(["./tests/results/test_sketch_dir"])
+        .spawn();
+    let mut cmd = Command::cargo_bin("sylph").unwrap();
+    let assert = cmd
+        .arg("sketch")
+        .arg("./test_files/e.coli-EC590.fasta")
+        .arg("./test_files/e.coli-K12.fasta")
+        .arg("./test_files/o157_reads.fastq")
+        .arg("./test_files/e.coli-W.fasta.gz")
+        .arg("-o")
+        .arg("./tests/results/test_sketch_dir/db")
+        .arg("-d")
+        .arg("./tests/results/test_sketch_dir")
+        .assert();
+    assert.success().code(0);
+
+    let mut cmd = Command::cargo_bin("sylph").unwrap();
+    let assert = cmd
+        .arg("profile")
+        .arg("./tests/results/test_sketch_dir/o157_reads.fastq.sylsp")
+        .arg("./tests/results/test_sketch_dir/db.syldb")
+        .assert();
+    assert.success().code(0);
+
+    let mut cmd = Command::cargo_bin("sylph").unwrap();
+    let assert = cmd
+        .arg("profile")
+        .arg("./tests/results/test_sketch_dir/o157_reads.fastq.sylsp")
+        .arg("./test_files/e.coli-EC590.fasta")
+        .assert();
+    assert.success().code(0);
+
+    let mut cmd = Command::cargo_bin("sylph").unwrap();
+    let assert = cmd
+        .arg("profile")
+        .arg("./test_files/o157_reads.fastq")
+        .arg("./test_files/e.coli-EC590.fasta")
+        .arg("-i")
+        .arg("-m")
+        .arg("90")
+        .assert();
+    assert.success().code(0);
+
+    let mut cmd= Command::cargo_bin("sylph").unwrap();
+    let assert = cmd
+        .arg("sketch")
+        .arg("-1")
+        .arg("./test_files/t1.fq")
+        .arg("-2")
+        .arg("./test_files/t2.fq")
+        .arg("-d")
+        .arg("./tests/results/test_sketch_dir")
+        .assert();
+    assert.success().code(0);
+    assert!(Path::new("./tests/results/test_sketch_dir/t1.fq.paired.sylsp").exists(), "Output file was not created");
+
+    fresh();
+    let mut cmd = Command::cargo_bin("sylph").unwrap();
+    let assert = cmd
+        .arg("sketch")
+        .arg("--sample-force")
+        .arg("./test_files/e.coli-EC590.fasta")
+        .arg("./test_files/o157_reads.fastq")
+        .arg("-o")
+        .arg("./tests/results/test_sketch_dir/db")
+        .arg("-d")
+        .arg("./tests/results/test_sketch_dir")
+        .assert();
+    assert.success().code(0);
+    assert!(Path::new("./tests/results/test_sketch_dir/e.coli-EC590.fasta.sylsp").exists(), "Output file was not created");
+    assert!(Path::new("./tests/results/test_sketch_dir/o157_reads.fastq.sylsp").exists(), "Output file was not created");
+    assert!(!Path::new("./tests/results/test_sketch_dir/db.syldb").exists(), "Output file was created");
+    fresh();
+
+    fresh();
+    let mut cmd = Command::cargo_bin("sylph").unwrap();
+    let assert = cmd
+        .arg("sketch")
+        .arg("--db-force")
+        .arg("./test_files/e.coli-EC590.fasta")
+        .arg("./test_files/o157_reads.fastq")
+        .arg("-o")
+        .arg("./tests/results/test_sketch_dir/db")
+        .arg("-d")
+        .arg("./tests/results/test_sketch_dir")
+        .assert();
+    assert.success().code(0);
+    assert!(!Path::new("./tests/results/test_sketch_dir/e.coli-EC590.fasta.sylsp").exists(), "Output file was created");
+    assert!(!Path::new("./tests/results/test_sketch_dir/o157_reads.fastq.sylsp").exists(), "Output file was created");
+    assert!(Path::new("./tests/results/test_sketch_dir/db.syldb").exists(), "Output file was not created");
+    fresh();
+}
+
+#[serial]
+#[test]
+fn test_profile_disabling(){
+    fresh();
+
+    let mut cmd = Command::cargo_bin("sylph").unwrap();
+    let assert = cmd
+        .arg("sketch")
+        .arg("--db-force")
+        .arg("./test_files/e.coli-EC590.fasta")
+        .arg("-o")
+        .arg("./tests/results/test_sketch_dir/db")
+        .arg("-d")
+        .arg("./tests/results/test_sketch_dir")
+        .arg("--disable-profiling")
+        .assert();
+    assert.success().code(0);
+
+    let mut output = Command::cargo_bin("sylph").unwrap();
+    let assert = output
+        .arg("profile")
+        .arg("./test_files/o157_reads.fastq")
+        .arg("./tests/results/test_sketch_dir/db.syldb")
+        .assert();
+    assert.failure().code(1);
+
+    let mut output = Command::cargo_bin("sylph").unwrap();
+    let assert = output
+        .arg("contain")
+        .arg("./test_files/o157_reads.fastq")
+        .arg("./tests/results/test_sketch_dir/db.syldb")
+        .assert();
+    assert.success().code(0);
+
+    fresh();
+}
+
+#[serial]
+#[test]
+fn test_sketch_fasta_fastq_concord(){
+    fresh();
+    let mut cmd = Command::cargo_bin("sylph").unwrap();
+    let assert = cmd
+        .arg("sketch")
+        .arg("./test_files/e.coli-EC590.fasta")
+        .arg("./test_files/o157_reads.fastq")
+        .arg("-o")
+        .arg("./tests/results/test_sketch_dir/db")
+        .arg("-d")
+        .arg("./tests/results/test_sketch_dir")
+        .assert();
+    assert.success().code(0);
+
+    let mut output = Command::cargo_bin("sylph").unwrap();
+    let out1 = output
+        .arg("profile")
+        .arg("./test_files/o157_reads.fastq")
+        .arg("./tests/results/test_sketch_dir/db.syldb")
+        .output()
+        .expect("Fail");
+
+    let mut output = Command::cargo_bin("sylph").unwrap();
+    let out2 = output
+        .arg("profile")
+        .arg("./test_files/o157_reads.fastq")
+        .arg("./test_files/e.coli-EC590.fasta")
+        .output()
+        .expect("Fail");
+
+    let mut output = Command::cargo_bin("sylph").unwrap();
+    let out3 = output
+        .arg("profile")
+        .arg("./tests/results/test_sketch_dir/o157_reads.fastq.sylsp")
+        .arg("./tests/results/test_sketch_dir/db.syldb")
+        .output()
+        .expect("Fail");
+
+    let stdout1 = str::from_utf8(&out1.stdout).expect("Output was not valid UTF-8");
+    let stdout2 = str::from_utf8(&out2.stdout).expect("Output was not valid UTF-8");
+    let stdout3 = str::from_utf8(&out3.stdout).expect("Output was not valid UTF-8");
+
+    assert!(stdout1 == stdout2);
+    assert!(stdout1 == stdout3);
+    assert!(stdout2 == stdout3);
+
+    fresh();
+}